Implement `FileGenerator`.
